fix: fmt

+++ src/Primitive.sol
@@ -15,7 +15,7 @@ contract Primitives {
 
     uint8 public u8 = 1;
     uint256 public u256 = 456;
-    uint public u = 123; // uint is an alias for uint256
+    uint256 public u = 123; // uint is an alias for uint256
 
     /*
     Negative numbers are allowed for int types.
@@ -27,7 +27,7 @@ contract Primitives {
     */
     int8 public i8 = -1;
     int256 public i256 = 456;
-    int public i = -123; // int is same as int256
+    int256 public i = -123; // int is same as int256
 
     // minimum and maximum of int
     int256 public minInt = type(int256).min;
@@ -43,8 +43,7 @@ contract Primitives {
     bytes10 d = 0x68656c6c6f776f726c64; // "helloworld" in hexadecimal
 
     // Declare a bytes32 array
-    bytes32 e =
-        0x68656c6c6f776f726c6420202020202020202020202020202020202020202020; // "helloworld" padded with spaces in hexadecimal
+    bytes32 e = 0x68656c6c6f776f726c6420202020202020202020202020202020202020202020; // "helloworld" padded with spaces in hexadecimal
 
     // Declare a dynamically-sized byte array
     bytes public c = "Hello, world!";
